move padding to own module as well
